[ZEPPELIN-5217]. Move %flink.cmd as a separated interpreter module
### What is this PR for?

Similar as ZEPPELIN-5213, just move %flink.cmd as separated interpreter module.

### What type of PR is it?
[Improvement]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-5217

### How should this be tested?
* CI pass

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #4038 from zjffdu/ZEPPELIN-5217 and squashes the following commits:

5e87345de [Jeff Zhang] [ZEPPELIN-5217]. Move %flink.cmd as a separated interpreter module

(cherry picked from commit 4cebdbf5aa2d04d37f6c729d22bd63f16f9a4694)
