+++ flink/interpreter/src/main/resources/interpreter-setting.json
@@ -276,17 +276,5 @@
       "completionKey": "TAB",
       "completionSupport": true
     }
-  },
-  {
-    "group": "flink",
-    "name": "cmd",
-    "className": "org.apache.zeppelin.flink.FlinkCmdInterpreter",
-    "properties": {
-    },
-    "editor": {
-      "language": "sh",
-      "editOnDblClick": false,
-      "completionSupport": false
-    }
   }
 ]
