+++ pom.xml
@@ -75,6 +75,7 @@
     <module>jdbc</module>
     <module>file</module>
     <module>flink</module>
+    <module>flink-cmd</module>
     <module>ignite</module>
     <module>influxdb</module>
     <module>kylin</module>
